requirements.txt
psycopg2-binary
pandas
statsmodels
streamlit
scikit-learn
plotly
^X

